+++ lib/ooxml_parser/xlsx_parser/xlsx_data/view_model/workbook/pivot_cache.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require_relative 'pivot_cache/pivot_cache_definition'
+
 module OoxmlParser
   # Class for parsing <pivotCache> tag
   class PivotCache < OOXMLDocumentObject
@@ -7,10 +9,8 @@ module OoxmlParser
     attr_reader :cache_id
     # @return [String] id of pivot cache
     attr_reader :id
-
-    def initialize(parent: nil)
-      @parent = parent
-    end
+    # @return [PivotCacheDefinition] parsed pivot cache definition
+    attr_reader :pivot_cache_definition
 
     # Parse Pivot Cache data
     # @param [Nokogiri::XML:Element] node with Pivot Cache data
@@ -24,7 +24,18 @@ module OoxmlParser
           @id = value.value.to_s
         end
       end
+      parse_pivot_cache_definition
       self
+    end
+
+    private
+
+    # @return [PivotCacheDefinition] pivot cache definition for current pivot cache
+    def parse_pivot_cache_definition
+      definition_file = root_object.relationships.target_by_id(id)
+      full_file_path = "#{OOXMLDocumentObject.path_to_folder}/xl/#{definition_file}"
+      @pivot_cache_definition = PivotCacheDefinition.new(parent: root_object)
+                                                    .parse(full_file_path)
     end
   end
 end